Group AES-GCM helpers together
